Rename BaseRecyclerAdapter to BaseListAdapter
Implement PagedListAdapter with BasePagedListAdapter
