+++ build.gradle
@@ -6,7 +6,9 @@ buildscript {
         recyclerViewVersion = '1.1.0'
         appcompatVersion = '1.1.0'
         archLifecycleVersion = '2.2.0'
-        ktxVersion = '1.2.0'
+        ktxVersion = '1.3.0'
+        pagingVersion = '1.0.0'
+
     }
     repositories {
         google()
@@ -14,7 +16,7 @@ buildscript {
 
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.6.1'
+        classpath 'com.android.tools.build:gradle:4.0.0'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
