+++ ui/src/main/java/com/kaloglu/library/ui/ViewExtensions.kt
@@ -9,10 +9,10 @@ import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.kaloglu.library.ui.adapter.BaseRecyclerAdapter
+import com.kaloglu.library.ui.adapter.BaseListAdapter
 
 @JvmOverloads
-fun <A : BaseRecyclerAdapter<*, *>> RecyclerView.setup(
+fun <A : BaseListAdapter<*, *>> RecyclerView.setup(
     adapter: A,
     layoutManager: RecyclerView.LayoutManager? = null,
     @DrawableRes seperatorDrawable: Int? = null
@@ -30,25 +30,25 @@ fun <A : BaseRecyclerAdapter<*, *>> RecyclerView.setup(
     return adapter
 }
 
-fun <A : BaseRecyclerAdapter<RI, *>, RI : RecyclerItem>
+fun <A : BaseListAdapter<RI, *>, RI : RecyclerItem>
         A.setItemClickListener(onClick: ((RI, Int) -> Unit)): A {
     this.onItemClick = onClick
     return this
 }
 
-fun <A : BaseRecyclerAdapter<RI, *>, RI : RecyclerItem>
+fun <A : BaseListAdapter<RI, *>, RI : RecyclerItem>
         A.setItemLongClickListener(onLongClick: ((RI, Int) -> Boolean)): A {
     this.onItemLongClick = onLongClick
     return this
 }
 
-fun <A : BaseRecyclerAdapter<RI, *>, RI : RecyclerItem>
+fun <A : BaseListAdapter<RI, *>, RI : RecyclerItem>
         A.setViewClickListener(onClick: ((RI, View, Int) -> Unit)): A {
     this.onViewClick = onClick
     return this
 }
 
-fun <A : BaseRecyclerAdapter<RI, *>, RI : RecyclerItem>
+fun <A : BaseListAdapter<RI, *>, RI : RecyclerItem>
         A.setViewLongClickListener(onLongClick: ((RI, View, Int) -> Boolean)): A {
     this.onViewLongClick = onLongClick
     return this
