+++ ui/src/main/java/com/kaloglu/library/ui/interfaces/ViewHolder.kt
@@ -6,7 +6,7 @@ interface ViewHolder<M> : LifecycleOwner, ViewHolderLifeCycle {
     fun bind(item: M)
 
     fun bind(
-        item: M,
+        item: M?,
         onItemClick: ((M, Int) -> Unit)? = null,
         onItemLongClick: ((M, Int) -> Boolean) = { _, _ -> false }
     )
